Pulled CLI parsing and signals detection into main execution file
